requirements.txt
ffmpeg-python
paddleocr
paddlepaddle
openai
pysrt